requirements.txt
# Core Agent Dependencies
anthropic==0.39.0
openai==1.54.0                    # Optional: for embeddings

# Async & HTTP
aiohttp==3.10.10
asyncio==3.4.3

# Data Sources
yfinance==0.2.48
fredapi==0.5.2
alpha-vantage==2.3.1

# Database
motor==3.6.0                      # Async MongoDB driver
pymongo>=4.9, <4.10              # MongoDB driver for synchronous operations

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Visualization & Dashboard
streamlit==1.40.1
plotly==5.24.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.2                  # Data validation
tenacity==9.0.0                   # Retry logic

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
jupyter==1.1.1